refactor: create 'visible' subfolder in `features`
