+++ src/app/catalog/_components/ObjectDetail/ObjectProperties/ObjectPropertiesWrapper.tsx
@@ -2,7 +2,7 @@
 import React, { useContext, useEffect, useRef, useState } from 'react'
 import Button from '@/ui/buttons/Button'
 import { Tab, TAB_TYPE, TabButton } from '@/app/catalog/_components/ObjectDetail/ObjectProperties/TabButton'
-import { AdaptiveContext } from '@/features/utility/adaptive'
+import { AdaptiveContext } from '@/features/visible/adaptive'
 
 interface Props {
   propsTabContent: React.ReactNode
