+++ src/components/QuickFilter/DetailFilterButton.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React, { useContext, useEffect, useState } from 'react'
-import { PopupContext } from '@/features/Popup'
+import { PopupContext } from '@/features/visible/Popup'
 import { useCategoryFilter } from '@/features/catalog/useCategoryFilter'
 
 interface Props {
