+++ src/layout/Header/CatalogMenu/components/CatalogMenuButton.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import React, { useContext } from 'react'
-import { toggleScroll } from '@/features/scroll'
+import { toggleScroll } from '@/features/visible/scroll'
 import { HEADER_MENUS, HeaderContext } from '@/layout/Header/Header.context'
 
 function CatalogMenuButton() {
