+++ src/layout/Header/components/SaleRentButton.tsx
@@ -1,6 +1,6 @@
 'use client'
 import React, { useContext } from 'react'
-import { hideScroll, showScroll } from '@/features/scroll'
+import { hideScroll, showScroll } from '@/features/visible/scroll'
 import { HeaderContext } from '@/layout/Header/Header.context'
 import { MenuType } from '@/globals/api'
