+++ src/ui/forms/CallForm.tsx
@@ -4,7 +4,7 @@ import { ConvertToCustomInputsMap, Form, FormImperativeRef, InputsMap } from '@/
 import React, { PropsWithChildren, useContext, useRef } from 'react'
 import { sendCallRequest } from '@/globals/api'
 import { INPUT_NAMES } from '@/globals/inputs/call-form'
-import { PopupContext } from '@/features/Popup'
+import { PopupContext } from '@/features/visible/Popup'
 
 type Inputs = ConvertToCustomInputsMap<typeof INPUT_NAMES>
